// src/app/master/client/shared/client.model.ts
export class ClientItem {
  id: number = 0;
  clientName: string = "";
  enabled: true;
  loginEnabled: false;
  description: string = "";
  clientAcronym: string = "";
  requestTypeId: number = 0;

  createdDate: string = '';
  createdBy: string = '';

  modifiedDate: string = '';
  modifiedBy: string = '';
  userName: string = '';
  userEmail: string = "";
}